updates lsp to remove mason dependancy
use nixpkgs to install instead, use lsp to configure only
